requirements.txt
torch
gdown
torchvision
seaborn
scikit-learn
pandas
pytest
flake8
simplejson
opencv-python
torchnet
tqdm
lmdb
pyyaml
gitpython
